nuScenes/requirements.txt
matplotlib
motmetrics
torch
numba
numpy
numpyencoder
nuscenes_devkit
opencv_python
pyquaternion
scipy
Shapely
tqdm
